Update README.md
It is impossible to overstate the importance of this book.

Michelle Alexander shows how slavery in the United States has not disappeared - it has just changed shape, into the mass incarceration of black men. 

The author, a former supreme court justice aid, meticulously describes and delineates the "War on Drugs" as being a system designed to 1) repress inner city lower income communities and effectively turn them into red zones, 2) place "criminals" under constant supervision and monitoring of the state, 3) prescribe harsh and unhelpful sentences for drug use and possession beyond the range of 'justice,' 4) criminalize poverty, 5) reward state and local drug enforcement agencies in the form of cash (federal subsidies for picking up and prosecuting drug criminals), and 6) effectively remove on third of the population from being able to vote in elections.