Add Cosmicomics
This is lesser-known Calvino, but actually my favorite of his works that I've read. It's just so weird! — a collection of stories about the deep history of the universe, featuring decidedly atypical protagonists (particles, formulae, evolving organisms, etc.) and a range of witty magical realism… Reads like vignettes from a strange alternate history of the cosmos…