Add Thus Spoke Zarathustra
This is one of my favorite books in general, and definitely my favorite philosophy book. I've read and re-read this book at pivotal moments in my life. Nietzsche is one of the first to spread the philosophy of "God is dead!" that now resembles Satanic beliefs, and is still a widely read book amongst Nihilists of all flavors. This book is not merely a critique on Christianity and organized religion; it is a manifesto of free will that people still read to this day. I highly recommend it as a mind-expanding book.